chore: lower RAM size from 1GB in places where 4GB was used (#40)
## What ❔

This PR lowers the RAM size for the simulation to 1GB where 4GB was used
before.

## Why ❔

#23 changed the memory layout and heap size making it possible to run
the simulation with 1GB of RAM instead of 4GB.
Tracing a simulation with 4GB of RAM required 12GB of CPU RAM just for
tracking the memory state, this change lowers this to 3GB, making
running multiple simulation in parallel more feasible.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted.