perf(gpu_prover): L2 cache persistence and multistreaming for production NTT sizes (#32)
## What ❔

Changes NTT internals to leverage L2 cache persistence, along with
2-stream ping pong approach to minimize tail effect.

As a bonus, i threw in 2-stream ping pong for making device columns sum
to zero, which shaves a few milliseconds by allowing the 1-block cleanup
kernel for each column to overlap with the main reduction kernel for the
next column.

## Why ❔

It looks like our production column lengths will be 2^22,23,24. Previous
experiments showed keeping a working set in the GPU's L2 cache between
kernel launches can significantly improve performance. This is trickier
for larger columns that might not fit in L2, but we can still manage it
for parts of the calculation, especially during the most NTT-heavy
phases (main domain evals->coset evals in stage 1 and stage 2).

## Is this a breaking change?
- [ ] Yes
- [x] No, prover should be faster out of the box

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted.