fix(gpu_prover): Avoid race condition in memory commitments (#56)
## What ❔

I encountered a race condition in the first pass memory commitments. It
manifested more often for faster GPUs and larger proofs. @robik75 had a
hunch, and after some sync-bisecting the present diff appears to be
enough to avoid the problem, even on H100 for full blocks of ethereum
transactions simulated in the ZKsync OS memory model (~hundreds of
millions of riscv cycles).

## Why ❔

Avoid the race condition. (Ideally we should be able to fix it without
syncing, but this can be a followup.)

## Is this a breaking change?
- [ ] Yes
- [x] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted.